+++ codex-rs/core/src/agent/control_tests.rs
@@ -6,20 +6,10 @@ use crate::agent::agent_status_from_event;
 use crate::config::AgentRoleConfig;
 use crate::config::Config;
 use crate::config::ConfigBuilder;
-use crate::config_loader::ConfigLayerEntry;
-use crate::config_loader::ConfigLayerStack;
-use crate::config_loader::ConfigLayerStackOrdering;
-use crate::config_loader::ConfigRequirements;
 use crate::config_loader::LoaderOverrides;
-use crate::config_loader::RequirementSource;
-use crate::config_loader::Sourced;
 use crate::contextual_user_message::SUBAGENT_NOTIFICATION_OPEN_TAG;
 use crate::features::Feature;
 use assert_matches::assert_matches;
-use codex_app_server_protocol::ConfigLayerSource;
-use codex_config::RequirementsExecPolicy;
-use codex_execpolicy::Decision;
-use codex_execpolicy::Policy;
 use codex_protocol::config_types::ModeKind;
 use codex_protocol::models::ContentItem;
 use codex_protocol::models::ResponseItem;
@@ -132,76 +122,6 @@ fn history_contains_text(history_items: &[ResponseItem], needle: &str) -> bool {
             ContentItem::InputImage { .. } => false,
         })
     })
-}
-
-#[tokio::test]
-async fn child_uses_parent_exec_policy_when_layer_stack_matches() {
-    let (_home, parent_config) = test_config().await;
-    let child_config = parent_config.clone();
-
-    assert!(child_uses_parent_exec_policy(&parent_config, &child_config));
-}
-
-#[tokio::test]
-async fn child_uses_parent_exec_policy_when_non_exec_policy_layers_differ() {
-    let (_home, parent_config) = test_config().await;
-    let mut child_config = parent_config.clone();
-    let mut layers: Vec<_> = child_config
-        .config_layer_stack
-        .get_layers(ConfigLayerStackOrdering::LowestPrecedenceFirst, true)
-        .into_iter()
-        .cloned()
-        .collect();
-    layers.push(ConfigLayerEntry::new(
-        ConfigLayerSource::SessionFlags,
-        TomlValue::Table(Default::default()),
-    ));
-    child_config.config_layer_stack = ConfigLayerStack::new(
-        layers,
-        child_config.config_layer_stack.requirements().clone(),
-        child_config.config_layer_stack.requirements_toml().clone(),
-    )
-    .expect("config layer stack");
-
-    assert!(child_uses_parent_exec_policy(&parent_config, &child_config));
-}
-
-#[tokio::test]
-async fn child_does_not_use_parent_exec_policy_when_requirements_exec_policy_differs() {
-    let (_home, parent_config) = test_config().await;
-    let mut child_config = parent_config.clone();
-    let mut requirements = ConfigRequirements {
-        exec_policy: child_config
-            .config_layer_stack
-            .requirements()
-            .exec_policy
-            .clone(),
-        ..ConfigRequirements::default()
-    };
-    let mut policy = Policy::empty();
-    policy
-        .add_prefix_rule(&["rm".to_string()], Decision::Forbidden)
-        .expect("add prefix rule");
-    requirements.exec_policy = Some(Sourced::new(
-        RequirementsExecPolicy::new(policy),
-        RequirementSource::Unknown,
-    ));
-    child_config.config_layer_stack = ConfigLayerStack::new(
-        child_config
-            .config_layer_stack
-            .get_layers(ConfigLayerStackOrdering::LowestPrecedenceFirst, true)
-            .into_iter()
-            .cloned()
-            .collect(),
-        requirements,
-        child_config.config_layer_stack.requirements_toml().clone(),
-    )
-    .expect("config layer stack");
-
-    assert!(!child_uses_parent_exec_policy(
-        &parent_config,
-        &child_config
-    ));
 }
 
 async fn wait_for_subagent_notification(parent_thread: &Arc<CodexThread>) -> bool {
